Updated UI as per  feedback
