+++ frontend/public/bg.jpg
