+++ frontend/public/login.html
@@ -1,0 +1,8 @@
+
+<header class="nav">
+  <nav class="navlinks">
+    <a href="#about">About</a>
+    <a href="#faqs">FAQs</a>
+    <a class="signin" href="login.html">Sign In</a>
+  </nav>
+</header>
