+++ frontend/src/Pages/Auth/LoginPage.js
@@ -1,7 +1,5 @@
 import React, { useState } from 'react';
 import './LoginPage.css';
-import logo from './logo.png';
-import bg from './bg.jpg';
 
 export default function LoginPage({ onLogin, onSignUpClick }) {
   const [email, setEmail] = useState('');
@@ -24,14 +22,14 @@ export default function LoginPage({ onLogin, onSignUpClick }) {
       </nav>
 
       <div className="login-right">
-        <img src={bg} alt="Login bg" className="bg-img" />
+        <img src="/bg.jpg" alt="Login bg" className="bg-img" />
         <div className="image-overlay"></div>
       </div>
 
       <div className="login-left">
         <div className="login-content">
           <div className="logo-section">
-            <img src={logo} alt="AutoAudit Logo" className="logo-img" />
+            <img src="/logo.png" alt="AutoAudit Logo" className="logo-img" />
             <h1>AutoAudit</h1>
             <p className="subtitle">Microsoft 365 Compliance Platform</p>
           </div>
