+++ frontend/src/Pages/Landing/AboutUs.css
@@ -51,6 +51,7 @@
   height: auto;
   margin-bottom: 30px;
   border-radius: 15px;
+  margin-left: 350px;
   box-shadow: 0 0 20px rgba(40, 224, 214, 0.3);
 }
